Simplify logpull code; pull into new package
We now allow the caller to pass a list of fields to fetch from the
Logpull endpoint, as well as an optional count paramter to limit the
number of returned results. Both of these are passed directly to the
Logpull endpoint and handled on Cloudflare's side.

Additionally, we now return an io.ReadCloser containing the raw NDJSON
log data from Cloudflare. After realizing that the encoding/json package
can decode NDJSON from an io.Reader, it became apparent that this would
afford the caller the most flexibility.

All API errors are now returned with a concrete type of HTTPError, which
captures the HTTP status code and response body.

The testing code is much simpler now as well: it turns out that the only
things we care about are that the API request is valid, that we return
the response body as received from Cloudflare on success, and that any
non-200 response is returned as an error which can be type-asserted to a
*HTTPError with the correct HTTP status code and response body.